FIX: Correct temperature units. Style improvments
The units of temperature were not being handled well because they were
inconsistent in Celsius/Kelvin, causing some values to be incorrect. An issue
was opened to let all of the unit conversions be handled by a package such as
pint.

Style was also improved by making several PEP-8 type changes and by converting
all of the `range(len(foo))` to `enumerate(foo)`. All of the instances required
`enumerate` because the `numpy` matrix needs to be indexed and also the looped
variable was used.