ENH: Separate plotting and calculating (closes #3)
Plotting operations have been moved out of the `System` class and into the module
`plot.py`. `System` now simply calculates and returns the 2d arrays that are used
in the plotting routines. The calculations should be more testable now because
they are directly returned from functions. Additionally, the import requirements
in the `system.py` module have been reduced to just `json`, `itertools`, and
`numpy`. The plot module is written with top level functions and these could
probably be decorated or the functions could be aliases that are set in some way
if support for other plotting packages is to be done in the future. The silver
and iron scripts are modified to reflect these changes.