TEST: Tests for grain boundary energy calculation
Tests were written for the grain boundary energy calculation. The tests compare
the values calculated for some calculations to a certain precision, so that any
changes to the grain boundary calculation will make the tests fail. Also some
minor error checking was implemented in the energy calculation. ValueError
exceptions will be raised if the temperature is below zero or if the
concentrations are below zero or above 1 or 0.5 for the grain boundary and
overall solute concentrations respectively.