Replace optimizer with scipy's fsolve. 2x speedup.
In addition to replacing the optimizer, the a grain size of 250nm is checked in
the beginning of the function, greatly improving the convergence times for
functions that would not converge. 250nm is chosen just as before, as it should
not be practical to calculate such grain sizes with this code.

Commit also fixes the '_vec' name in the solute_can_stabilize function. Need
tests for these to determine errors like this are not introduced.