Swich optimize_grain_size from fsolve to brentq
Changing provided a minor speed boost of about 20%. More importantly, brentq
allows for explicitly setting bounds, where 1 to 100 was chosen for the grain
size. This prevents needing to do work-arounds in fsolve to prevent a negative
grain size ever being passed (which throws an exception). The maximum grain size
was lowered from 250 to 100, further reinforcing the bounds of the problem.
Theoretically, reducing it might speed up convergence, which is the reason why it
was lowered. Again, beyond 100 nm becomes less useful and applicable to this
problem domain.