FIX: Add bounds argument to optimize_grain_size
Fixes #1. For now, this allows for the customization of bounds, so that the
bounds can be adapted around the root. For Ag-W (example in issue #1) These could
be 10nm and 100nm instead of 1. Since there are some that are stable below these,
this change is not enforced and (1, 100) will be the default bounds. It's possible
to programatically update if you're finding that the solute can stabilize (via the
function of the same name) by doing a simple binary search on the bounds to find
ones where the shape of the energy is more desirable (again, see issue #1 for the
energy shapes).